Move the UTF-16 unit oracle into its test
casefold_utf16_units, and the UTF-8 validator standing behind it, have
no production caller: the escape decision already covers ill-formed
UTF-8 through its byte >= 0x80 rule, and the unit budgets are enforced
by static assert. The host lane's budget check is the only consumer.

Move both there. Inside the codec, that check compares the production
arithmetic against itself. In the test the counter is a second
implementation, so a codec change that breaks the unit budget shows up
as a disagreement between the two instead of being mirrored into both
sides at once.
